Fix IO Blocking
Apparently `readline()` is a blocking call and will not return until it
reads a whole line. Changing this to `read(1)` means that we read and
print one byte at a time, and we also ignore the io-blocking put in
place when the process waits for input. This fixes the issues with input
prompts showing up late.

Some more discussion of the issue here:
https://stackoverflow.com/questions/375427/non-blocking-read-on-a-subprocess-pipe-in-python

Also added flush=True to the print command to force flushing of the
buffer.