Fix bug in plan check when directory path is a prefix of another (#26)
* fix bug in plan check when directory path is a prefix of another

the plan check would sometimes pick up the wrong directories when using --modified-only if a modified directory
path was a substring of the root directory even if the directory was not a child. For example:

root_directory /foo/bar
modified_directory /foo/bar2

in the above case, /foo/bar2 is not a child of the root directory so we shouldn't run plan on it even if
some files under it have been modified. /foo/bar2 is a substring of /foo/bar though and this would
cause the plan_check to run plan on that directory incorrectly

* add a comment