Optionally Allow Less Printing for tf_apply
Added a `--print-only-changes` option to tf_apply, that will change the
behavior of tf_apply such that it only prints Terraform output for
directories that actually changed.

Additionally, tf_apply will now not apply a directory if the plan comes
back clean for it.

Additionally, tf_apply will now run init and the desired command in one
operation, which should make applies overall faster and reduce some
needless printing that was going on.