BugFix: DynamoDB Unlock
Two fixes here... The DynamoDB unlock functionality would fail when
called from repos that are not called terraform-config, so broke out the
function for figuring out the backend state path so it can be reused.

I also added additional error handling where we delete the lock if the
state file is deleted.