AT-4667 Fix Plan Check for Terraform 0.15
Apparently, Terraform 0.15 no longer allows -lock=false for the init
command, so we cannot pass it as an argument in plan_check anymore. So
instead only pass it when we are running the actual plan command.

-lock=false apparently never did anything for init in the first place,
so it should be safe to remove this entirely without trying to do some
kind of version check.