Fix: Ignore State Warnings
The latest version of Terraform will evaluate data sources as print out any diffs that it finds separately, with a note that the Terraform plan might affect those resources.

This change will ignore the output of the plan command in the plan_check until we get to the actual changes that Terraform intends to make, which might include those IAM resources that it warned about above. If the resources are actually included there, then they will be changed and it would be appropriate to flag the plan. If they aren't actually included there, the Terraform engine has decided that the config is sufficiently close that no changes are necessary and we shouldn't flag the plan.